Update Defines.cpp
This change is in response to this issue, https://github.com/cuberite/cuberite/issues/5441. In the event that AddFaceDirection is called with an invalid eBlockFace the function will simply return the position is was called with, rather than crashing the server. Another solution to this is outlined by madmaxoft in a comment on the original issue, but this is beyond my abilities to implement at the moment, so consider this a temporary solution if that would work better for the purposes of the project.